ai-editor-2.0-full-stack-application/workers/requirements.txt
# Redis Queue for background jobs
rq==2.1.0
redis==5.2.1

# API clients
anthropic==0.49.0
openai==1.58.1
google-generativeai==0.8.3
pyairtable==3.0.2
requests==2.32.3

# Database
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
pillow==11.1.0
cloudinary==1.42.0

# For image processing
aiohttp==3.11.11

# HTTP Trigger Service
flask==3.1.0

# RSS Feed Parsing
feedparser==6.0.11

# Web Scraping / Content Extraction
firecrawl-py==1.15.0
